requirements.txt
# =============================================================================
# requirements.txt — Python dependencies for MODCAN-GNN Pipeline
# Môi trường: Python 3.10+ | GPU: CUDA 12.1
# Cài đặt trong Docker (đã được xử lý bởi Dockerfile).
# Để cài cục bộ (không GPU): pip install -r requirements.txt
# =============================================================================

# ── Core Scientific Stack ──────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
scipy==1.13.0
scikit-learn==1.4.2
joblib==1.4.0

# ── Deep Learning & GNN ───────────────────────────────────────────────────────
# NOTE: Cài PyTorch và PyG từ wheel tương ứng với CUDA version
# torch==2.3.0+cu121 -- từ https://download.pytorch.org/whl/cu121
# torch-geometric==2.5.3

# ── Graph Analysis ────────────────────────────────────────────────────────────
networkx==3.3

# ── Bioinformatics ────────────────────────────────────────────────────────────
biopython==1.83
pyranges==0.0.129

# ── Visualization ─────────────────────────────────────────────────────────────
matplotlib==3.9.0
seaborn==0.13.2
lifelines==0.29.0
umap-learn==0.5.6
plotly==5.22.0

# ── Data I/O ───────────────────────────────────────────────────────────────────
h5py==3.11.0
pyarrow==16.0.0
openpyxl==3.1.3
requests==2.32.2

# ── ML Utilities ──────────────────────────────────────────────────────────────
imbalanced-learn==0.12.2
shap==0.45.0
optuna==3.6.1

# ── Configuration & Utilities ─────────────────────────────────────────────────
pyyaml==6.0.1
omegaconf==2.3.0
tqdm==4.66.4
loguru==0.7.2
click==8.1.7

# ── Jupyter ───────────────────────────────────────────────────────────────────
jupyterlab==4.2.0
ipywidgets==8.1.3
